Use basic auth for dashboard view, but not badges
This commit makes application of basic auth (if active) more selective.
Instead of forcing basic auth on all endpoints, it is now applied as a
decorator for the main dashboard views. This allows us to serve the
status badges without basic auth while still being able to protect the
dashboard pages.