Link pipeline paths in nested nav to filter dashboard
The nested nav titles are now linked so they will load the dashboard
with filtered for the value of the link. E.g. If at dashboard root, and
there are two pipeline collections, 'europe/croatia' and
'america/mexico', clicking on 'europe/croatia' will reload the dashboard
with the content filtered to just display croatia's pipelines.