Pipeline status badge endpoints for server (#158)
* Fix pipeline status badge

Some previous status manager refactoring broke the badge end point. I've
also changed it from redirecting to the shields.io image, to proxying
it.

* If no pipeline_id found, still return a badge

If no pipeline_id is found, return a badge with 'not found' written in
it, rather than returning a 404. Knowing that a pipeline is potentially
missing is still useful status.

* Pipeline collection status badge

A status badge at the /badge/collection/<pipeline path> endpoint.

This returns a pipeline collection summary badge, e.g.

pipelines | 22 succeeded, 14 failed, 2 invalid, 1 running

Pipeline path will filter the pipeline ids it is reporting on.

* Make pipeline status text lowercase to fit conventions

* Update readme with status badge details.

* Use basic auth for dashboard view, but not badges

This commit makes application of basic auth (if active) more selective.
Instead of forcing basic auth on all endpoints, it is now applied as a
decorator for the main dashboard views. This allows us to serve the
status badges without basic auth while still being able to protect the
dashboard pages.

* Removing the refresh view as it's not used and looks like a DoS risk.

* Add note to README about basic auth and the badges