Filter dashboard by pipeline collection (#159)
* Link dashboard header to url_for main index page

* Allow dashboard filtering by pipeline path

The main view function now accepts a pipeline path to filter the
pipeline results it displays. This allows users to focus only on
pipelines and pipeline collection they are interested in. E.g.

http://localhost:5000/europe/croatia

This will display all pipelines where the id starts with
./europe/croatia.

* Link pipeline paths in nested nav to filter dashboard

The nested nav titles are now linked so they will load the dashboard
with filtered for the value of the link. E.g. If at dashboard root, and
there are two pipeline collections, 'europe/croatia' and
'america/mexico', clicking on 'europe/croatia' will reload the dashboard
with the content filtered to just display croatia's pipelines.